+++ bld/gui/h/gui.h
@@ -706,25 +706,25 @@ extern bool GUIFillRectRGB( gui_window *wnd, gui_rect *rect, gui_rgb rgb );
 extern bool GUIDrawRectRGB( gui_window *wnd, gui_rect *rect, gui_rgb rgb );
 extern bool GUIDrawLineRGB( gui_window *wnd, gui_point *start, gui_point *end,
                          gui_line_styles style, gui_ord thickness, gui_rgb rgb );
-extern void GUIDrawText( gui_window *wnd, const char *text, int length,
+extern void GUIDrawText( gui_window *wnd, const char *text, size_t length,
                          gui_ord row, gui_ord indent, gui_attr attr );
-extern void GUIDrawTextPos( gui_window *wnd, const char *text, int length,
+extern void GUIDrawTextPos( gui_window *wnd, const char *text, size_t length,
                             gui_coord *pos, gui_attr attr );
-extern void GUIDrawTextExtent( gui_window *wnd, const char *text, int length,
+extern void GUIDrawTextExtent( gui_window *wnd, const char *text, size_t length,
                                gui_ord row, gui_ord indent, gui_attr attr,
                                gui_ord extentx );
-extern void GUIDrawTextExtentPos( gui_window *wnd, const char *text, int length,
+extern void GUIDrawTextExtentPos( gui_window *wnd, const char *text, size_t length,
                                gui_coord *pos, gui_attr attr, gui_ord extentx );
-extern void GUIDrawTextRGB( gui_window *wnd, char *text, int length,
+extern void GUIDrawTextRGB( gui_window *wnd, char *text, size_t length,
                             gui_ord row, gui_ord indent,
                             gui_rgb fore, gui_rgb back );
-extern void GUIDrawTextPosRGB( gui_window *wnd, char *text, int length,
+extern void GUIDrawTextPosRGB( gui_window *wnd, char *text, size_t length,
                                gui_coord *pos, gui_rgb fore, gui_rgb back );
-extern void GUIDrawTextExtentRGB( gui_window *wnd, char *text, int length,
+extern void GUIDrawTextExtentRGB( gui_window *wnd, char *text, size_t length,
                                   gui_ord row, gui_ord indent,
                                   gui_rgb fore, gui_rgb back,
                                   gui_ord extentx );
-extern void GUIDrawTextExtentPosRGB( gui_window *wnd, char *text, int length,
+extern void GUIDrawTextExtentPosRGB( gui_window *wnd, char *text, size_t length,
                                      gui_coord *pos,
                                      gui_rgb fore, gui_rgb back,
                                      gui_ord extentx );
@@ -738,15 +738,15 @@ extern bool GUIDrawBarGroup( gui_window *wnd, gui_ord row, gui_ord start,
 /* Text Functions */
 
 extern bool GUISetWindowText( gui_window * wnd, const char * data );
-extern int GUIGetWindowTextLength( gui_window *wnd );
-extern int GUIGetWindowText( gui_window *wnd, char *data, int max_length );
+extern size_t GUIGetWindowTextLength( gui_window *wnd );
+extern size_t GUIGetWindowText( gui_window *wnd, char *data, size_t max_length );
 extern gui_ord GUIGetRow( gui_window * wnd, gui_point *pos );
 extern gui_ord GUIGetCol( gui_window * wnd, char *text, gui_point *pos );
 extern gui_ord GUIGetStringPos( gui_window *wnd, gui_ord indent,
                                 char * string, int mouse_x );
-extern gui_ord GUIGetExtentX( gui_window *wnd, const char * text, int length );
+extern gui_ord GUIGetExtentX( gui_window *wnd, const char * text, size_t length );
 extern gui_ord GUIGetExtentY( gui_window *wnd, const char * text );
-extern gui_ord GUIGetControlExtentX( gui_window * wnd, unsigned id, const char * text, int length );
+extern gui_ord GUIGetControlExtentX( gui_window * wnd, unsigned id, const char * text, size_t length );
 extern gui_ord GUIGetControlExtentY( gui_window * wnd, unsigned id, const char * text );
 extern void GUIGetTextMetrics( gui_window *wnd, gui_text_metrics *metrics );
 extern void GUIGetDlgTextMetrics( gui_text_metrics *metrics );
